requirements.txt
easydict
einops
fastapi
gdown
gradio<=5.4.0
h5py
huggingface_hub
imageio
Jinja2
kornia
loguru
matplotlib<3.9
numpy~=1.24
omegaconf
onnxruntime
opencv-contrib-python
opencv-python
packaging
pandas
plotly
poselib
protobuf
psutil
pycolmap==0.6.1
pydantic==2.10.6
pynvml
pytlsd
pytorch-lightning==1.4.9
PyYAML
requests
scikit-image
scikit-learn
scipy
seaborn
shapely
safetensors
tensorboardX==2.6.1
torch==2.2.2
torchmetrics==0.6.0
torchvision==0.17.2
tqdm
uvicorn
yacs
